Adding LocalStoreTestHelper for NO_CHANGE event
